+++ pkg/core/diagnostics.go
@@ -1,18 +1,10 @@
 package core
 
 import (
-	"strings"
-
 	"github.com/cloudquery/cloudquery/pkg/plugin"
 	"github.com/cloudquery/cq-provider-sdk/provider/diag"
 	"github.com/spf13/viper"
 )
-
-type DiagnosticsSummary struct {
-	Total      int            `json:"total,omitempty"`
-	ByType     map[string]int `json:"by_type,omitempty"`
-	BySeverity map[string]int `json:"by_severity,omitempty"`
-}
 
 type SentryDiagnostic struct {
 	diag.Diagnostic
@@ -38,32 +30,6 @@ func (d *SentryDiagnostic) Redacted() diag.Diagnostic {
 
 func (d *SentryDiagnostic) IsSentryDiagnostic() (bool, map[string]string, bool) {
 	return true, d.Tags, d.Ignore
-}
-
-func SummarizeDiagnostics(diags diag.Diagnostics) DiagnosticsSummary {
-	summary := DiagnosticsSummary{
-		Total:      0,
-		ByType:     make(map[string]int),
-		BySeverity: make(map[string]int),
-	}
-	for _, d := range diags {
-		summary.Total++
-		severity := strings.ToLower(d.Severity().String())
-		dtype := strings.ToLower(d.Type().String())
-
-		if count, ok := summary.BySeverity[severity]; ok {
-			summary.BySeverity[severity] = count + 1
-		} else {
-			summary.BySeverity[severity] = 1
-		}
-
-		if count, ok := summary.ByType[dtype]; ok {
-			summary.ByType[dtype] = count + 1
-		} else {
-			summary.ByType[dtype] = 1
-		}
-	}
-	return summary
 }
 
 func convertToConfigureDiags(diags diag.Diagnostics) diag.Diagnostics {
